+++ backend/src/baserow/contrib/database/views/signals.py
@@ -1,4 +1,9 @@
-from django.dispatch import Signal
+from django.dispatch import Signal, receiver
+
+from baserow.contrib.database.fields.signals import field_deleted
+from baserow.contrib.database.fields.models import FileField
+
+from .models import GalleryView
 
 
 view_created = Signal()
@@ -14,3 +19,11 @@ view_sort_created = Signal()
 view_sort_updated = Signal()
 view_sort_deleted = Signal()
 view_field_options_updated = Signal()
+
+
+@receiver(field_deleted)
+def field_updated(sender, field, **kwargs):
+    if isinstance(field, FileField):
+        GalleryView.objects.filter(card_cover_image_field_id=field.id).update(
+            card_cover_image_field_id=None
+        )